refactor(arch): relocate convention types from markdownflavor to internal/convention
Hoist Convention, RulePreset, and Flavor data shapes plus
ParseFlavor / Lookup / Names out of the markdownflavor rule
into a new internal/convention package. internal/config now
imports internal/convention instead of reaching into a rule
package, restoring the layering direction recorded in the
architecture hub.

The markdownflavor rule keeps its rule.Rule impl and the
Feature/support table that maps flavor → which features it
accepts; r.Flavor is now convention.Flavor and Supports is a
free function in the rule package taking a convention.Flavor.

Adds TestConfigDoesNotImportRules to guard the new direction:
it parses every non-test file under internal/config/ and fails
if any import path contains internal/rules/.

Closes plan/155.
