+++ internal/rules/markdownflavor/features.go
@@ -1,93 +1,12 @@
 // Package markdownflavor implements MDS034, which validates Markdown
 // against a declared target flavor (commonmark, gfm, goldmark,
 // pandoc, phpextra, multimarkdown, myst, or any) and flags syntax
-// the target renderer will not understand.
+// the target renderer will not understand. The Flavor type itself,
+// along with the convention bundles that select it, lives in
+// internal/convention; this package consumes those data shapes.
 package markdownflavor
 
-// Flavor identifies a target Markdown flavor.
-type Flavor int
-
-// Flavor constants. The zero value is intentionally invalid so that
-// unparsed settings are caught.
-const (
-	flavorInvalid Flavor = iota
-	FlavorCommonMark
-	FlavorGFM
-	FlavorGoldmark
-	// FlavorAny accepts every tracked feature. Useful when the
-	// document is destined for an unknown or permissive renderer and
-	// the user wants to disable flavor reporting without disabling
-	// the rule.
-	FlavorAny
-	// FlavorPandoc is Pandoc's default markdown dialect. Accepts
-	// GFM's four features plus footnotes, definition lists, heading
-	// IDs, superscript, subscript, math block, and inline math;
-	// rejects abbreviations (a non-default Pandoc extension).
-	FlavorPandoc
-	// FlavorPHPExtra is PHP Markdown Extra. Accepts tables,
-	// footnotes, definition lists, heading IDs, and abbreviations;
-	// rejects GFM's task lists, strikethrough, bare-URL autolinks,
-	// and every math / sub/superscript feature.
-	FlavorPHPExtra
-	// FlavorMultiMarkdown extends PHP Markdown Extra with math
-	// block and inline math. Like PHP Extra, rejects GFM task lists,
-	// strikethrough, bare-URL autolinks, and sub/superscript.
-	FlavorMultiMarkdown
-	// FlavorMyST is the MyST flavor used by the Sphinx documentation
-	// toolchain. Accepts tables, strikethrough, footnotes,
-	// definition lists, heading IDs, math block, and inline math;
-	// rejects GFM task lists, bare-URL autolinks, sub/superscript,
-	// and abbreviations.
-	FlavorMyST
-)
-
-// String returns the canonical lowercase name of the flavor.
-func (f Flavor) String() string {
-	switch f {
-	case FlavorCommonMark:
-		return "commonmark"
-	case FlavorGFM:
-		return "gfm"
-	case FlavorGoldmark:
-		return "goldmark"
-	case FlavorAny:
-		return "any"
-	case FlavorPandoc:
-		return "pandoc"
-	case FlavorPHPExtra:
-		return "phpextra"
-	case FlavorMultiMarkdown:
-		return "multimarkdown"
-	case FlavorMyST:
-		return "myst"
-	}
-	return ""
-}
-
-// ParseFlavor converts a config string into a Flavor. The match is
-// case-sensitive to reject typos like "GFM" that would otherwise
-// silently validate against the wrong flavor.
-func ParseFlavor(s string) (Flavor, bool) {
-	switch s {
-	case "commonmark":
-		return FlavorCommonMark, true
-	case "gfm":
-		return FlavorGFM, true
-	case "goldmark":
-		return FlavorGoldmark, true
-	case "any":
-		return FlavorAny, true
-	case "pandoc":
-		return FlavorPandoc, true
-	case "phpextra":
-		return FlavorPHPExtra, true
-	case "multimarkdown":
-		return FlavorMultiMarkdown, true
-	case "myst":
-		return FlavorMyST, true
-	}
-	return 0, false
-}
+import "github.com/jeduden/mdsmith/internal/convention"
 
 // Feature identifies one Markdown syntax feature whose support varies
 // across flavors.
@@ -167,23 +86,23 @@ func (f Feature) Name() string {
 // lists, strikethrough, and bare-URL autolinks. The goldmark flavor
 // further adds heading IDs. Pandoc, PHP Markdown Extra, MultiMarkdown,
 // and MyST each pick a different combination of the optional
-// features; FlavorAny is handled specially in Supports.
-var support = map[Flavor]map[Feature]bool{
-	FlavorGFM: {
+// features; convention.FlavorAny is handled specially in Supports.
+var support = map[convention.Flavor]map[Feature]bool{
+	convention.FlavorGFM: {
 		FeatureTables:           true,
 		FeatureTaskLists:        true,
 		FeatureStrikethrough:    true,
 		FeatureBareURLAutolinks: true,
 		FeatureGitHubAlerts:     true,
 	},
-	FlavorGoldmark: {
+	convention.FlavorGoldmark: {
 		FeatureTables:           true,
 		FeatureTaskLists:        true,
 		FeatureStrikethrough:    true,
 		FeatureBareURLAutolinks: true,
 		FeatureHeadingIDs:       true,
 	},
-	FlavorPandoc: {
+	convention.FlavorPandoc: {
 		FeatureTables:           true,
 		FeatureTaskLists:        true,
 		FeatureStrikethrough:    true,
@@ -196,14 +115,14 @@ var support = map[Flavor]map[Feature]bool{
 		FeatureMathBlock:        true,
 		FeatureMathInline:       true,
 	},
-	FlavorPHPExtra: {
+	convention.FlavorPHPExtra: {
 		FeatureTables:          true,
 		FeatureFootnotes:       true,
 		FeatureDefinitionLists: true,
 		FeatureHeadingIDs:      true,
 		FeatureAbbreviations:   true,
 	},
-	FlavorMultiMarkdown: {
+	convention.FlavorMultiMarkdown: {
 		FeatureTables:          true,
 		FeatureFootnotes:       true,
 		FeatureDefinitionLists: true,
@@ -212,7 +131,7 @@ var support = map[Flavor]map[Feature]bool{
 		FeatureMathBlock:       true,
 		FeatureMathInline:      true,
 	},
-	FlavorMyST: {
+	convention.FlavorMyST: {
 		FeatureTables:          true,
 		FeatureStrikethrough:   true,
 		FeatureFootnotes:       true,
@@ -224,10 +143,10 @@ var support = map[Flavor]map[Feature]bool{
 }
 
 // Supports reports whether the flavor accepts the given feature.
-// FlavorAny accepts every feature; other flavors consult the
-// support table.
-func (f Flavor) Supports(feat Feature) bool {
-	if f == FlavorAny {
+// convention.FlavorAny accepts every feature; other flavors consult
+// the support table.
+func Supports(f convention.Flavor, feat Feature) bool {
+	if f == convention.FlavorAny {
 		return true
 	}
 	return support[f][feat]
