+++ internal/rules/markdownflavor/features_test.go
@@ -5,68 +5,24 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/jeduden/mdsmith/internal/convention"
 )
-
-func TestParseFlavor(t *testing.T) {
-	tests := []struct {
-		in   string
-		want Flavor
-		ok   bool
-	}{
-		{"commonmark", FlavorCommonMark, true},
-		{"gfm", FlavorGFM, true},
-		{"goldmark", FlavorGoldmark, true},
-		{"any", FlavorAny, true},
-		{"pandoc", FlavorPandoc, true},
-		{"phpextra", FlavorPHPExtra, true},
-		{"multimarkdown", FlavorMultiMarkdown, true},
-		{"myst", FlavorMyST, true},
-		{"GFM", 0, false},
-		{"", 0, false},
-		{"markdown", 0, false},
-	}
-	for _, tc := range tests {
-		t.Run(tc.in, func(t *testing.T) {
-			got, ok := ParseFlavor(tc.in)
-			assert.Equal(t, tc.ok, ok)
-			if tc.ok {
-				assert.Equal(t, tc.want, got)
-			}
-		})
-	}
-}
-
-func TestFlavorString(t *testing.T) {
-	assert.Equal(t, "commonmark", FlavorCommonMark.String())
-	assert.Equal(t, "gfm", FlavorGFM.String())
-	assert.Equal(t, "goldmark", FlavorGoldmark.String())
-	assert.Equal(t, "any", FlavorAny.String())
-	assert.Equal(t, "pandoc", FlavorPandoc.String())
-	assert.Equal(t, "phpextra", FlavorPHPExtra.String())
-	assert.Equal(t, "multimarkdown", FlavorMultiMarkdown.String())
-	assert.Equal(t, "myst", FlavorMyST.String())
-}
 
 // assertSupports checks every feature in the supported set is
 // accepted by flavor and every feature not in that set is rejected.
-func assertSupports(t *testing.T, f Flavor, supported ...Feature) {
+func assertSupports(t *testing.T, f convention.Flavor, supported ...Feature) {
 	t.Helper()
 	want := map[Feature]bool{}
 	for _, feat := range supported {
 		want[feat] = true
 	}
 	for _, feat := range AllFeatures() {
-		got := f.Supports(feat)
+		got := Supports(f, feat)
 		assert.Equal(t, want[feat], got,
 			"flavor %s feature %s: want=%v got=%v",
 			f.String(), feat.Name(), want[feat], got)
 	}
-}
-
-func TestFlavorStringUnknownIsEmpty(t *testing.T) {
-	var zero Flavor
-	assert.Equal(t, "", zero.String())
-	assert.Equal(t, "", Flavor(999).String())
 }
 
 func TestFeatureNameUnknownIsEmpty(t *testing.T) {
@@ -74,27 +30,27 @@ func TestFeatureNameUnknownIsEmpty(t *testing.T) {
 }
 
 func TestFeatureSupportCommonMark(t *testing.T) {
-	assertSupports(t, FlavorCommonMark)
+	assertSupports(t, convention.FlavorCommonMark)
 }
 
 func TestFeatureSupportGFM(t *testing.T) {
-	assertSupports(t, FlavorGFM,
+	assertSupports(t, convention.FlavorGFM,
 		FeatureTables, FeatureTaskLists, FeatureStrikethrough,
 		FeatureBareURLAutolinks, FeatureGitHubAlerts)
 }
 
 func TestFeatureSupportGoldmark(t *testing.T) {
-	assertSupports(t, FlavorGoldmark,
+	assertSupports(t, convention.FlavorGoldmark,
 		FeatureTables, FeatureTaskLists, FeatureStrikethrough,
 		FeatureBareURLAutolinks, FeatureHeadingIDs)
 }
 
 func TestFeatureSupportAny(t *testing.T) {
-	assertSupports(t, FlavorAny, AllFeatures()...)
+	assertSupports(t, convention.FlavorAny, AllFeatures()...)
 }
 
 func TestFeatureSupportPandoc(t *testing.T) {
-	assertSupports(t, FlavorPandoc,
+	assertSupports(t, convention.FlavorPandoc,
 		FeatureTables, FeatureTaskLists, FeatureStrikethrough,
 		FeatureBareURLAutolinks, FeatureFootnotes, FeatureDefinitionLists,
 		FeatureHeadingIDs, FeatureSuperscript, FeatureSubscript,
@@ -102,20 +58,20 @@ func TestFeatureSupportPandoc(t *testing.T) {
 }
 
 func TestFeatureSupportPHPExtra(t *testing.T) {
-	assertSupports(t, FlavorPHPExtra,
+	assertSupports(t, convention.FlavorPHPExtra,
 		FeatureTables, FeatureFootnotes, FeatureDefinitionLists,
 		FeatureHeadingIDs, FeatureAbbreviations)
 }
 
 func TestFeatureSupportMultiMarkdown(t *testing.T) {
-	assertSupports(t, FlavorMultiMarkdown,
+	assertSupports(t, convention.FlavorMultiMarkdown,
 		FeatureTables, FeatureFootnotes, FeatureDefinitionLists,
 		FeatureHeadingIDs, FeatureAbbreviations,
 		FeatureMathBlock, FeatureMathInline)
 }
 
 func TestFeatureSupportMyST(t *testing.T) {
-	assertSupports(t, FlavorMyST,
+	assertSupports(t, convention.FlavorMyST,
 		FeatureTables, FeatureStrikethrough, FeatureFootnotes,
 		FeatureDefinitionLists, FeatureHeadingIDs,
 		FeatureMathBlock, FeatureMathInline)
